+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # v0.0.3
 
 * added pluggable pathpoint value databases
-
+* `Section` is now a `Future`